src/zhihu/RAG/requirements.txt
pdfminer.six
chromadb
nltk
elasticsearch7
sentence_transformers